refactor : types/index.ts 분리
index.ts의 타입들을 data.ts와 props.ts로 분리후 각 의존 컴포넌트에 변경사항 적용
