+++ src/components/search/SearchForm.tsx
@@ -3,13 +3,8 @@ import styled from "styled-components";
 import magnifyImg from '../../assets/images/magnify.png';
 import axios from "axios";
 import {makeFoodImgURL, makeFoodRawMaterialURL} from "../../utils";
-import {
-    FoodImageData,
-    FoodImageResponse, FoodMaterialData,
-    FoodMaterialResponse,
-    SearchFormProps,
-    SearchResult
-} from "../../type";
+import {FoodImageData, FoodImageResponse, FoodMaterialData, FoodMaterialResponse, SearchResult} from "../../type/data";
+import {SearchFormProps} from "../../type/props";
 
 const Wrapper = styled.div`
   display: flex;
