+++ src/components/search/SearchList.tsx
@@ -2,7 +2,8 @@ import React, {useEffect} from "react";
 import styled from "styled-components";
 import SearchItem from "./SearchItem";
 import {Skeleton} from "antd";
-import {SearchListProps, SearchResult, SearchScrollBoxProps} from "../../type";
+import {SearchListProps, SearchScrollBoxProps} from "../../type/props";
+import {SearchResult} from "../../type/data";
 
 const Wrapper = styled.div`
   display: flex;
