+++ src/pages/SearchPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PageTemplate from "../layout/PageTemplate";
 import SearchContent from "../components/search/SearchContent";
-import {SearchPageProps} from "../type";
+import {SearchPageProps} from "../type/props";
 
 function SearchPage({searchResults,handleSearchResultsChange}:SearchPageProps) {
     return (
